+++ data_handlers/serverbound_packets/11/237.cjs
@@ -1,0 +1,43 @@
+const {Socket, World} = require('../../../data_structures.cjs')
+const dataReader = require('../../data_reader.cjs')
+const packetWriter = require('../../clientbound_packets/packet_writer.cjs')
+const utils = require('../../../utils/utils.cjs')
+
+var packetID = 11
+var packetIdentifier = "Player Position"
+
+/** 
+ * @param {World} world 
+ * @param {Socket} socket 
+ * @param {Buffer} data 
+ */
+function ReadPacket(world, socket, data) {
+    var splitIndex = data.length - 34
+
+    socket.log(`SERVERBOUND --> ${packetID} "${packetIdentifier}" / ${data.length} bytes`, false)
+
+    if (splitIndex >= 0) {
+        var position = {}
+        position.x = dataReader.readDouble(socket, data, 1)
+        position.y = dataReader.readDouble(socket, data, position.x.nextPos)
+        position.stance = dataReader.readDouble(socket, data, position.y.nextPos)
+        position.z = dataReader.readDouble(socket, data, position.stance.nextPos)
+        
+        var onGround = dataReader.readBool(socket, data, position.z.nextPos)
+
+        if (socket.disconnect == "" && !socket.thisPlayer.tick.teleportSelf && socket.thisPlayer.allowMovement) {
+            var newPosition = {x: position.x.value, y: position.y.value, z: position.z.value}
+
+            if ((position.x.value == 8.5 && position.y.value == 65 && position.z.value == 8.5) || position.y.value < 1) {
+                socket.thisPlayer.tick.teleportSelf = true
+                return
+            }
+
+            utils.player.set.Position(world, socket.thisPlayer, newPosition)
+        }
+    }
+    
+    return splitIndex
+}
+
+module.exports = {ReadPacket}